Adds Install Script for PDFMiner
This is a simple little script for downloading and unpacking the pdfminer
library for Python. It still should be moved to
lib/python2.7/site-packages in your local virtual environment (depending
on what directory that is installed in). I might handle this differently
later, but I couldn't find it in pip, and I'm not sure if I want to pull
it from Git through requirements.txt. So, for now it's in a BASh script.