Script for Processing PDFs by Page
This adds a script that contains a function that processes PDF files
page-by-page. It takes in a single PDF file as an argument, and returns
a generator that yields out the text of each page of the document.

The goal of this is to pass the pages into a datastore--likely a
document-oriented database, so that the text can be indexed and searched,
and maybe so that the document can be returned as a search result.